feat(explorer wp): inject content into baked page
