+++ db/wpdb.ts
@@ -508,7 +508,7 @@ export async function getRelatedCharts(
     `)
 }
 
-export async function getBlock(id: number): Promise<string | undefined> {
+export async function getBlockContent(id: number): Promise<string | undefined> {
     const WP_GRAPHQL_ENDPOINT = `${WORDPRESS_URL}/wp/graphql`
     const query = `
     query getBlock($id: ID!) {
